fix rapids require spark-avro issues
